pyproject.toml
[tool.poetry]
name = "coinbase_dca"
version = "0.1.0"
description = ""
authors = ["Alex Melville"]
# readme = "README.md"
# license = "BSD"
packages = [{ include = "coinbase_dca" }]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
coinbase-advanced-py = "^1.1.3"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
main = "coinbase_dca.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
